Arreglo en las paginas
