+++ src/App.css
@@ -1,0 +1,3 @@
+.App {
+      background-image: rgb(199, 11, 11);
+}
