+++ src/components/CheckoutPage.js
@@ -3,8 +3,9 @@ import { Box } from '@mui/system';
 import Grid from '@mui/material/Grid';
 import Typography from '@mui/material/Typography';
 import CheckoutCard  from "./CheckoutCard";
-//import products from "../product-data";
+import Button from '@mui/material/Button';
 import Total from "./Total";
+import { Link } from "react-router-dom";
 
 
 
@@ -13,7 +14,7 @@ const CheckoutPage = () => {
       // Hooks
 
       const [cart, setCart] = useState([]);
-      
+
       // Fetch
       
       const getDataFromAPI = async() => {
@@ -38,7 +39,7 @@ const CheckoutPage = () => {
     function FormRow(){
       return (
             <React.Fragment>
-             {cart.length === 0 ? "No hay productos en tu carrito" : cart.map((item) => (
+             {cart.map((item) => (
               <Grid item xs={12} sm={8} md={6} lg={4}>
                  <CheckoutCard key={item.id} product={item} deleteProductCart={deleteProductCart} />
               </Grid>
@@ -47,13 +48,14 @@ const CheckoutPage = () => {
       );  
     }
 
+  if(cart.length > 0){ 
     return (
         <Box sx={{ flexGrow: 1 , padding: "2rem"}}>
          <Grid container spacing={3}>
-            <Grid item xs={12}>
+            <Grid item xs={12}>           
               <Typography align='center' gutterBottom variant='h4'>
                 Mi carrito
-              </Typography>
+              </Typography>          
             </Grid>
             <Grid item xs={12} sm={8} md={9} container spacing={2}>
                 <FormRow />
@@ -67,6 +69,25 @@ const CheckoutPage = () => {
          </Grid>
         </Box>
     );
+  }
+  else{
+    return (
+      <Box sx={{ flexGrow: 1 , padding: "2rem"}}>
+       <Grid container spacing={2}>
+          <Grid item xs={12}>           
+            <Typography align='left' gutterBottom variant='h4'>
+              No hay productos en tu carrito
+            </Typography>         
+          </Grid>
+          <Grid item xs={12}> 
+             <Link to= "/">
+              <Button marginTop= "2rem" variant="contained" color="success" > Volver a la tienda  </Button>
+             </Link>
+          </Grid> 
+       </Grid>
+      </Box>
+   );
+  }
 };
 
 export default CheckoutPage;
