+++ src/components/Total.js
@@ -11,6 +11,9 @@ const total = cart.reduce((acc, item) => {
     return acc.plus(itemPrice.times(item.quantity)); 
 }, new Decimal(0)); 
 
+const compra = cart;
+ 
+if(cart.length > 0) {
     return (
         <Box sx={{ display: "flex", flexDirection: "column", justifyContent: "center", alignItems: "center", height: "20vh" }}>
             <h5>Total item: {cart.reduce((acc, item) => acc + item.quantity, 0) }</h5>
@@ -18,6 +21,7 @@ const total = cart.reduce((acc, item) => {
                 <Button marginTop= "2rem" variant="contained" color="primary"> Check out </Button>          
         </Box>
     )
+ }
 }
 
 export default Total
